Restructure hardware checks to enrich node info

+++ miner/src/operations/structs/node.rs
@@ -1,6 +1,6 @@
 use serde::{Deserialize, Serialize};
 
-#[derive(Serialize, Deserialize)]
+#[derive(Serialize, Deserialize, Debug)]
 pub struct ComputeSpecs {
     pub gpu: Option<GpuSpecs>,
     pub cpu: Option<CpuSpecs>,
@@ -8,23 +8,23 @@ pub struct ComputeSpecs {
     pub storage_gb: Option<u32>,
 }
 
-#[derive(Serialize, Deserialize)]
+#[derive(Serialize, Deserialize, Debug)]
 pub struct GpuSpecs {
     pub count: Option<u32>,
     pub model: Option<String>,
     pub memory_mb: Option<u32>,
 }
 
-#[derive(Serialize, Deserialize)]
+#[derive(Serialize, Deserialize, Debug)]
 pub struct CpuSpecs {
     pub cores: Option<u32>,
     pub model: Option<String>,
 }
 
-#[derive(Serialize, Deserialize)]
+#[derive(Serialize, Deserialize, Debug)]
 pub struct NodeConfig {
     pub ip_address: String,
-    pub port: u32,
+    pub port: u16,
     pub compute_specs: Option<ComputeSpecs>,
     pub compute_pool_id: u64,
 }
